detect/file-data: consolidate matching code
